api/requirements.txt
flask==3.1.0
gunicorn
pyodbc==4.0.26
azure-identity==1.23.1
azure-keyvault-secrets==4.10.0
